Create data_exporter_beta_compatible.gs
Updated file to be compatible with the new Google Ads beta. Changed lines 180 - 194, so we first find label IDs (using GQL with escaped regex to deal with the '$') and then use the label IDs to in the label selector. This should be compatible both with and without the beta. I've tested both with and without, and all seems to work as expected.